Make cleanUpData function more robust in design
This function was not managing to handle a larger data set and
generated a new list that skipped some of the `null` venue check-ins
but also failed to leave out quite a lot of them. Not entirely clear
on why this happens. But:

This change is more robust, because even with a similar limitation,
an incomplete new list is an improvement empty venues.